+++ Core/DigestFile.cs
@@ -117,6 +117,35 @@
                               already in a using block of its own, thus ensuring
                               that the stream is properly closed and releases
                               its unmanaged resources including its file handle.
+
+	2026/03/31 9.0.386 DAG    This update extends output formats to include
+                              Base64 encoded strings, with or without a prefix
+                              that identifies the hashing algorithm used to
+                              generate the digest. The prefix is derived from
+                              the name of the hashing algorithm, and is followed
+                              by an underscore character. For example, a 16-byte
+                              MD5 digest would be represented by a string of 26
+                              characters, beginning with "MD5_", and a 64-byte
+                              SHA-512 digest would be represented by a string of
+                              92 characters, beginning with "SHA512_". The 
+                              output format is specified by an optional
+                              parameter of type OuutputFormat, which is an
+                              enumeration that defines the supported output
+                              formats for the digest. The method uses a switch
+                              statement to determine how to format the digest
+                              based on the specified output format.
+
+                              Though the immediate need for the additional
+                              formats is confined to SHA-384, all hashing
+                              algorithms are updated to support the new output
+                              formats, and the new output formats are added to
+                              the documentation for all hashing algorithms, and
+                              the file I/O is moved into a private nethod, which
+                              is called by all hashing algorithms, eliminating
+                              the file I/O code from the hashing algorithm
+                              methods, and centralizing it in a single method
+                              that can be easily maintained and updated as
+                              needed, and making the code slightly more compact.
     ============================================================================
 */
 
@@ -126,7 +155,6 @@ using System.Collections.Generic;
 using System.IO;
 using System.Security.Cryptography;
 
-using WizardWrx;
 
 namespace WizardWrx.Cryptography
 {
@@ -144,16 +172,86 @@ namespace WizardWrx.Cryptography
     /// </summary>
     public static class DigestFile
     {
-        const int BUFSIZE = 8192;
+		/// <summary>
+		/// Members of this public enumeration specify the output format of the message digest.
+		/// </summary>
+		public enum OuutputFormat
+		{
+			/// <summary>
+			/// Output the message digest as a string of hexadecimal digits. The
+			/// number of hexadecimal digits is twice the number of bytes in the
+			/// digest, since each byte is represented by two hexadecimal
+			/// digits. For example, the 16-byte MD5 digest is represented by a
+			/// string of 32 hexadecimal digits, and the 64-byte SHA-512 digest
+			/// is represented by a string of 128 hexadecimal digits.
+			/// </summary>
+			HexadecimalDigits,
+
+			/// <summary>
+			/// Output the message as a Base64 encoded string. The number of
+			/// characters in the Base64 string is approximately 4/3 the number
+			/// of bytes in the digest, since each group of 3 bytes is
+			/// represented by 4 Base64 characters. For example, the 16-byte MD5
+			/// digest is represented by a string of 22 Base64 characters, and
+			/// the 64-byte SHA-512 digest is represented by a string of 88
+			/// Base64 characters.
+			/// </summary>
+			Base64String,
+
+			/// <summary>
+			/// A Base64-encoded string that includes a prefix that identifies
+			/// the hashing algorithm used to generate the digest. The prefix is
+			/// derived from the name of the hashing algorithm, and is followed
+			/// by an underscore character. For example, the 16-byte MD5 digest
+			/// is represented by a string of 26 characters, beginning with
+			/// "MD5_", and the 64-byte SHA-512 digest is represented by a
+			/// string of 92 characters, beginning with "SHA512_".
+			/// </summary>
+			PrefixedBase64String
+		}   // public enum OutputFormat
+
+
+		/// <summary>
+		/// This private enumeration is used to identify the hashing algorithm
+		/// used to generate the digest, so that the appropriate prefix can be
+		/// added when the output format is PrefixedBase64String.
+		/// </summary>
+		enum DigestType
+		{
+			DIGEST_TYPE_MD5,
+			DIGEST_TYPE_SHA1,
+			DIGEST_TYPE_SHA256,
+			DIGEST_TYPE_SHA384,
+			DIGEST_TYPE_SHA512
+		}   // enum DIGEST_TYPE
+
+
+		/// <summary>
+		/// This dictionary maps each DigestType to its corresponding string
+		/// prefix, which is used when the output format is
+		/// PrefixedBase64String. The prefixes are derived from the names of the
+		/// hashing algorithms, and are followed by an underscore character. For
+		/// example, the prefix for DigestType DIGEST_TYPE_MD5 is "MD5_", and
+		/// the prefix for DigestType DIGEST_TYPE_SHA512 is "SHA512_".
+		/// </summary>
+		static readonly Dictionary<DigestType , string> s_dctHashAlgorithmPrefixes = new Dictionary<DigestType , string> ( )
+		{
+			{ DigestType.DIGEST_TYPE_MD5 , @"MD5" } ,
+			{ DigestType.DIGEST_TYPE_SHA1 , @"SHA1" } ,
+			{ DigestType.DIGEST_TYPE_SHA256 , @"SHA256" } ,
+			{ DigestType.DIGEST_TYPE_SHA384 , @"SHA384" } ,
+			{ DigestType.DIGEST_TYPE_SHA512 , @"SHA512" }
+		};
+
 
 		/// <summary>
 		/// Given the name of a file, return its MD5 message digest as a 32 
 		/// character string of hexadecimal digits.
 		/// </summary>
 		/// <param name="pstrFileName">
-		/// <para>
-		/// String containing the name of the file to be digested.
-		/// </para>
+		/// This string represents the name of the file to process, which may be
+		/// a relative or absolute path. The file must exist, and the caller
+		/// must have permission to read it, or an exception will be thrown.
 		/// <para>
 		/// iMPORTANT: Since files are read in binary, they are loaded directly
 		/// into the hash algorithm as byte arrays. This means that the digest
@@ -163,6 +261,20 @@ namespace WizardWrx.Cryptography
 		/// characters yields a different byte stream than the byte stream that
 		/// came from reading the file in binary mode.
 		/// </para>
+		/// </param>
+		/// <param name="penmOutputFormat">
+		/// This optional parameter specifies the output format of the message
+		/// digest. If omitted, the default is OuutputFormat.HexadecimalDigits,
+		/// which outputs the digest as a string of hexadecimal digits. The
+		/// other options are OuutputFormat.Base64String, which outputs the
+		/// digest as a Base64 encoded string, and
+		/// OuutputFormat.PrefixedBase64String, which outputs a Base64 encoded
+		/// string that includes a prefix that identifies the hashing algorithm
+		/// used to generate the digest. The prefix is derived from the name of
+		/// the hashing algorithm, and is followed by an underscore character.
+		/// For example, the 16-byte MD5 digest is represented by a string of 26
+		/// characters, beginning with "MD5_", and the 64-byte SHA-512 digest is
+		/// represented by a string of 92 characters, beginning with "SHA512_".
 		/// </param>
 		/// <returns>
 		/// The message digest, consisting of a string of 32 hexadecimal
@@ -175,17 +287,6 @@ namespace WizardWrx.Cryptography
 		/// Since the internal hashing implementations expect byte arrays, the
 		/// input string must be converted. The Encoding.Default.GetBytes method
 		/// is called upon to convert the string into a byte array.
-		///</para>
-		///<para>
-		/// A slightly modified version of the Rivest code, written in ANSI C,
-		/// is the engine in my MD5WIN stand-alone program and my MD5Digest
-		/// Windows Dynamic Link Library.
-		///</para>
-		///<para>
-		/// However, this function uses a MD5CryptoServiceProvider object, which
-		/// provides a managed interface to the Cryptographic Service Provider
-		/// in the host's installation of Microsoft Windows. I chose this over
-		/// the 100% managed implementation for two reasons.
 		///</para>
 		///<list>
 		///<item>
@@ -200,16 +301,18 @@ namespace WizardWrx.Cryptography
 		///</list>
 		///</remarks>
 		[Obsolete ( "This algorithm is classified as broken and unsafe. Use SHA256Hash, SHA384Hash, or SHA512Hash." )]
-        public static string MD5Hash ( string pstrFileName )
-        {
-            using ( FileStream fsInput = new FileStream ( pstrFileName , FileMode.Open , FileAccess.Read , FileShare.Read , BUFSIZE , FileOptions.SequentialScan | FileOptions.Asynchronous ) )
-            {
-                using ( MD5 md5Hasher = new MD5CryptoServiceProvider ( ) )
-                {
-                    byte [ ] bytInputData = md5Hasher.ComputeHash ( fsInput );
-                    return Core.ByteArrayFormatters.ByteArrayToHexDigitString ( bytInputData );
-				}   // using ( MD5 md5Hasher = new MD5CryptoServiceProvider ( ) )
-			}   // using ( FileStream fsInput = new FileStream ( pstrFileName , FileMode.Open , FileAccess.Read , FileShare.Read , BUFSIZE , FileOptions.SequentialScan | FileOptions.Asynchronous ) )
+        public static string MD5Hash ( string pstrFileName , OuutputFormat penmOutputFormat = OuutputFormat.HexadecimalDigits )
+		{
+			byte [ ] abytFileContents = GetFileContent ( pstrFileName );
+
+			using ( MD5 sha = MD5.Create ( ) )
+			{
+				byte [ ] abytDigest = sha.ComputeHash ( abytFileContents );
+				return FormatDigestPerOutputFormat (
+					abytDigest ,
+					penmOutputFormat ,
+					DigestType.DIGEST_TYPE_MD5 );
+			}   // using ( MD5 sha = MD5.Create ( ) )
 		}   // public static string MD5Hash
 
 
@@ -218,9 +321,9 @@ namespace WizardWrx.Cryptography
 		/// character string of hexadecimal digits.
 		/// </summary>
 		/// <param name="pstrFileName">
-		/// <para>
-		/// String containing the name of the file to be digested.
-		/// </para>
+		/// This string represents the name of the file to process, which may be
+		/// a relative or absolute path. The file must exist, and the caller
+		/// must have permission to read it, or an exception will be thrown.
 		/// <para>
 		/// iMPORTANT: Since files are read in binary, they are loaded directly
 		/// into the hash algorithm as byte arrays. This means that the digest
@@ -231,21 +334,37 @@ namespace WizardWrx.Cryptography
 		/// came from reading the file in binary mode.
 		/// </para>
 		/// </param>
+		/// <param name="penmOutputFormat">
+		/// This optional parameter specifies the output format of the message
+		/// digest. If omitted, the default is OuutputFormat.HexadecimalDigits,
+		/// which outputs the digest as a string of hexadecimal digits. The
+		/// other options are OuutputFormat.Base64String, which outputs the
+		/// digest as a Base64 encoded string, and
+		/// OuutputFormat.PrefixedBase64String, which outputs a Base64 encoded
+		/// string that includes a prefix that identifies the hashing algorithm
+		/// used to generate the digest. The prefix is derived from the name of
+		/// the hashing algorithm, and is followed by an underscore character.
+		/// For example, the 16-byte MD5 digest is represented by a string of 26
+		/// characters, beginning with "MD5_", and the 64-byte SHA-512 digest is
+		/// represented by a string of 92 characters, beginning with "SHA512_".
+		/// </param>
 		/// <returns>
 		/// The message digest, consisting of a string of 40 hexadecimal
 		/// characters.
 		/// </returns>
 		[Obsolete ( "This algorithm is classified as broken and unsafe. Use SHA256Hash, SHA384Hash, or SHA512Hash." )]
-        public static string SHA1Hash ( string pstrFileName )
-        {
-            using ( FileStream fsInput = new FileStream ( pstrFileName , FileMode.Open , FileAccess.Read , FileShare.Read , BUFSIZE , FileOptions.SequentialScan | FileOptions.Asynchronous ) )
-            {
-                using ( SHA1 SHA1Hasher = new SHA1CryptoServiceProvider ( ) )
-                {
-                    byte [ ] bytInputData = SHA1Hasher.ComputeHash ( fsInput );
-                    return Core.ByteArrayFormatters.ByteArrayToHexDigitString ( bytInputData );
-				}   // using ( SHA1 SHA1Hasher = new SHA1CryptoServiceProvider ( ) )
-			}   // using ( FileStream fsInput = new FileStream ( pstrFileName , FileMode.Open , FileAccess.Read , FileShare.Read , BUFSIZE , FileOptions.SequentialScan | FileOptions.Asynchronous ) )
+        public static string SHA1Hash ( string pstrFileName , OuutputFormat penmOutputFormat = OuutputFormat.HexadecimalDigits )
+		{
+			byte [ ] abytFileContents = GetFileContent ( pstrFileName );
+
+			using ( SHA1 sha = SHA1.Create ( ) )
+			{
+				byte [ ] abytDigest = sha.ComputeHash ( abytFileContents );
+				return FormatDigestPerOutputFormat (
+					abytDigest ,
+					penmOutputFormat ,
+					DigestType.DIGEST_TYPE_SHA1 );
+			}   // using ( SHA1 sha = SHA1.Create ( ) )
 		}   // public static string SHA1Hash
 
 
@@ -254,9 +373,9 @@ namespace WizardWrx.Cryptography
 		/// character string of hexadecimal digits.
 		/// </summary>
 		/// <param name="pstrFileName">
-		/// <para>
-		/// String containing the name of the file to be digested.
-		/// </para>
+		/// This string represents the name of the file to process, which may be
+		/// a relative or absolute path. The file must exist, and the caller
+		/// must have permission to read it, or an exception will be thrown.
 		/// <para>
 		/// iMPORTANT: Since files are read in binary, they are loaded directly
 		/// into the hash algorithm as byte arrays. This means that the digest
@@ -267,20 +386,36 @@ namespace WizardWrx.Cryptography
 		/// came from reading the file in binary mode.
 		/// </para>
 		/// </param>
+		/// <param name="penmOutputFormat">
+		/// This optional parameter specifies the output format of the message
+		/// digest. If omitted, the default is OuutputFormat.HexadecimalDigits,
+		/// which outputs the digest as a string of hexadecimal digits. The
+		/// other options are OuutputFormat.Base64String, which outputs the
+		/// digest as a Base64 encoded string, and
+		/// OuutputFormat.PrefixedBase64String, which outputs a Base64 encoded
+		/// string that includes a prefix that identifies the hashing algorithm
+		/// used to generate the digest. The prefix is derived from the name of
+		/// the hashing algorithm, and is followed by an underscore character.
+		/// For example, the 16-byte MD5 digest is represented by a string of 26
+		/// characters, beginning with "MD5_", and the 64-byte SHA-512 digest is
+		/// represented by a string of 92 characters, beginning with "SHA512_".
+		/// </param>
 		/// <returns>
 		/// The message digest, consisting of a string of 64 hexadecimal
 		/// characters.
 		/// </returns>
-		public static string SHA256Hash ( string pstrFileName )
-        {
-            using ( FileStream fsInput = new FileStream ( pstrFileName , FileMode.Open , FileAccess.Read , FileShare.Read , BUFSIZE , FileOptions.SequentialScan | FileOptions.Asynchronous ) )
-            {
-                using ( SHA256 SHA256Hasher = new SHA256Managed ( ) )
-                {
-                    byte [ ] bytInputData = SHA256Hasher.ComputeHash ( fsInput );
-                    return Core.ByteArrayFormatters.ByteArrayToHexDigitString ( bytInputData );
-				}   // using ( SHA256 SHA256Hasher = new SHA256Managed ( ) )
-			}   // using ( FileStream fsInput = new FileStream ( pstrFileName , FileMode.Open , FileAccess.Read , FileShare.Read , BUFSIZE , FileOptions.SequentialScan | FileOptions.Asynchronous ) )
+		public static string SHA256Hash ( string pstrFileName , OuutputFormat penmOutputFormat = OuutputFormat.HexadecimalDigits )
+		{
+			byte [ ] abytFileContents = GetFileContent ( pstrFileName );
+
+			using ( SHA256 sha = SHA256.Create ( ) )
+			{
+				byte [ ] abytDigest = sha.ComputeHash ( abytFileContents );
+				return FormatDigestPerOutputFormat (
+					abytDigest ,
+					penmOutputFormat ,
+					DigestType.DIGEST_TYPE_SHA256 );
+			}   // using ( SHA256 sha = SHA256.Create ( ) )
 		}   // public static string SHA256Hash
 
 
@@ -289,9 +424,9 @@ namespace WizardWrx.Cryptography
 		/// character string of hexadecimal digits.
 		/// </summary>
 		/// <param name="pstrFileName">
-		/// <para>
-		/// String containing the name of the file to be digested.
-		/// </para>
+		/// This string represents the name of the file to process, which may be
+		/// a relative or absolute path. The file must exist, and the caller
+		/// must have permission to read it, or an exception will be thrown.
 		/// <para>
 		/// iMPORTANT: Since files are read in binary, they are loaded directly
 		/// into the hash algorithm as byte arrays. This means that the digest
@@ -302,21 +437,38 @@ namespace WizardWrx.Cryptography
 		/// came from reading the file in binary mode.
 		/// </para>
 		/// </param>
+		/// <param name="penmOutputFormat">
+		/// This optional parameter specifies the output format of the message
+		/// digest. If omitted, the default is OuutputFormat.HexadecimalDigits,
+		/// which outputs the digest as a string of hexadecimal digits. The
+		/// other options are OuutputFormat.Base64String, which outputs the
+		/// digest as a Base64 encoded string, and
+		/// OuutputFormat.PrefixedBase64String, which outputs a Base64 encoded
+		/// string that includes a prefix that identifies the hashing algorithm
+		/// used to generate the digest. The prefix is derived from the name of
+		/// the hashing algorithm, and is followed by an underscore character.
+		/// For example, the 16-byte MD5 digest is represented by a string of 26
+		/// characters, beginning with "MD5_", and the 64-byte SHA-512 digest is
+		/// represented by a string of 92 characters, beginning with "SHA512_".
+		/// </param>
 		/// <returns>
 		/// The message digest, consisting of a string of 64 hexadecimal
 		/// characters.
 		/// </returns>
-		public static string SHA384Hash ( string pstrFileName )
-        {
-            using ( FileStream fsInput = new FileStream ( pstrFileName , FileMode.Open , FileAccess.Read , FileShare.Read , BUFSIZE , FileOptions.SequentialScan | FileOptions.Asynchronous ) )
-            {
-                using ( SHA384 SHA384Hasher = new SHA384Managed ( ) )
-                {
-                    byte [ ] bytInputData = SHA384Hasher.ComputeHash ( fsInput );
-                    return Core.ByteArrayFormatters.ByteArrayToHexDigitString ( bytInputData );
-				}   // using ( SHA384 SHA384Hasher = new SHA384Managed ( ) )
-			}   // using ( FileStream fsInput = new FileStream ( pstrFileName , FileMode.Open , FileAccess.Read , FileShare.Read , BUFSIZE , FileOptions.SequentialScan | FileOptions.Asynchronous ) )
+		public static string SHA384Hash ( string pstrFileName , OuutputFormat penmOutputFormat = OuutputFormat.HexadecimalDigits )
+		{
+			byte [ ] abytFileContents = GetFileContent ( pstrFileName );
+
+			using ( SHA384 sha = SHA384.Create ( ) )
+			{
+				byte [ ] abytDigest = sha.ComputeHash ( abytFileContents );
+				return FormatDigestPerOutputFormat (
+					abytDigest ,
+					penmOutputFormat ,
+					DigestType.DIGEST_TYPE_SHA384 );
+			}   // using ( SHA384 sha = SHA384.Create ( ) )
 		}   // public static string SHA384Hash
+
 
 
 		/// <summary>
@@ -324,9 +476,9 @@ namespace WizardWrx.Cryptography
 		/// character string of hexadecimal digits.
 		/// </summary>
 		/// <param name="pstrFileName">
-		/// <para>
-		/// String containing the name of the file to be digested.
-		/// </para>
+		/// This string represents the name of the file to process, which may be
+		/// a relative or absolute path. The file must exist, and the caller
+		/// must have permission to read it, or an exception will be thrown.
 		/// <para>
 		/// iMPORTANT: Since files are read in binary, they are loaded directly
 		/// into the hash algorithm as byte arrays. This means that the digest
@@ -337,20 +489,205 @@ namespace WizardWrx.Cryptography
 		/// came from reading the file in binary mode.
 		/// </para>
 		/// </param>
+		/// <param name="penmOutputFormat">
+		/// This optional parameter specifies the output format of the message
+		/// digest. If omitted, the default is OuutputFormat.HexadecimalDigits,
+		/// which outputs the digest as a string of hexadecimal digits. The
+		/// other options are OuutputFormat.Base64String, which outputs the
+		/// digest as a Base64 encoded string, and
+		/// OuutputFormat.PrefixedBase64String, which outputs a Base64 encoded
+		/// string that includes a prefix that identifies the hashing algorithm
+		/// used to generate the digest. The prefix is derived from the name of
+		/// the hashing algorithm, and is followed by an underscore character.
+		/// For example, the 16-byte MD5 digest is represented by a string of 26
+		/// characters, beginning with "MD5_", and the 64-byte SHA-512 digest is
+		/// represented by a string of 92 characters, beginning with "SHA512_".
+		/// </param>
 		/// <returns>
 		/// The message digest, consisting of a string of 64 hexadecimal
 		/// characters.
 		/// </returns>
-		public static string SHA512Hash ( string pstrFileName )
-        {
-            using ( FileStream fsInput = new FileStream ( pstrFileName , FileMode.Open , FileAccess.Read , FileShare.Read , BUFSIZE , FileOptions.SequentialScan | FileOptions.Asynchronous ) )
-            {
-                using ( SHA512 SHA512Hasher = new SHA512Managed ( ) )
-                {
-                    byte [ ] bytInputData = SHA512Hasher.ComputeHash ( fsInput );
-                    return Core.ByteArrayFormatters.ByteArrayToHexDigitString ( bytInputData );
-				}   //  using ( SHA512 SHA512Hasher = new SHA512Managed ( ) )
-			}   // using ( FileStream fsInput = new FileStream ( pstrFileName , FileMode.Open , FileAccess.Read , FileShare.Read , BUFSIZE , FileOptions.SequentialScan | FileOptions.Asynchronous ) )
+		public static string SHA512Hash ( string pstrFileName , OuutputFormat penmOutputFormat = OuutputFormat.HexadecimalDigits )
+		{
+			byte [ ] abytFileContents = GetFileContent ( pstrFileName );
+
+			using ( SHA512 sha = SHA512.Create ( ) )
+			{
+				byte [ ] abytDigest = sha.ComputeHash ( abytFileContents );
+				return FormatDigestPerOutputFormat (
+					abytDigest ,
+					penmOutputFormat ,
+					DigestType.DIGEST_TYPE_SHA512 );
+			}   // using ( SHA512 sha = SHA512.Create ( ) )
 		}   // public static string SHA512Hash
-    }   // class public static class
+
+
+		/// <summary>
+		/// This private method formats the message digest according to the
+		/// specified output format. The digest is provided as a byte array, and
+		/// the output format is specified by the OuutputFormat enumeration. The
+		/// method also takes a DigestType parameter to determine the
+		/// appropriate prefix when the output format is PrefixedBase64String.
+		/// The method uses a switch statement to determine how to format the
+		/// digest based on the specified output format. If the output format is
+		/// HexadecimalDigits, it converts the byte array to a string of
+		/// hexadecimal digits using the ByteArrayToHexDigitString method. If
+		/// the output format is Base64String, it converts the byte array to a
+		/// Base64 encoded string using the Convert.ToBase64String method. If
+		/// the output format is PrefixedBase64String, it retrieves the
+		/// appropriate prefix from the s_dctHashAlgorithmPrefixes dictionary
+		/// using the DigestType parameter, and then constructs the output
+		/// string by concatenating the prefix, an underscore character, and the
+		/// Base64 encoded string of the digest. If the output format does not
+		/// match any of the specified cases, it defaults to returning the
+		/// digest as a string of hexadecimal digits.
+		/// </summary>
+		/// <param name="abytDigest">
+		/// This byte array represents the message digest that was computed by
+		/// the hashing algorithm. The contents of this byte array will be
+		/// formatted according to the specified output format and returned as a
+		/// string. The length of the byte array depends on the hashing
+		/// algorithm used to compute the digest. For example, the MD5 algorithm
+		/// produces a 16-byte digest, the SHA-1 algorithm produces a 20-byte
+		/// digest, the SHA-256 algorithm produces a 32-byte digest, the SHA-384
+		/// algorithm produces a 48-byte digest, and the SHA-512 algorithm
+		/// produces a 64-byte digest. The contents of the byte array will be
+		/// the raw binary representation of the message digest, and it will be
+		/// formatted into a human-readable string based on the specified output
+		/// format.
+		/// </param>
+		/// <param name="penmOutputFormat">
+		/// This parameter specifies the output format of the message digest. It
+		/// is of type OuutputFormat, which is an enumeration that defines the
+		/// possible output formats for the digest. The method will use this
+		/// parameter to determine how to format the byte array of the digest
+		/// into a string. The possible values for this parameter are:
+		/// <list type="number">
+		/// <item>
+		/// OuutputFormat.HexadecimalDigits produces a string of hexadecimal
+		/// digits, where each byte in the digest is represented by two
+		/// hexadecimal characters. For example, a 16-byte digest would be
+		/// represented by a string of 32 hexadecimal characters.
+		/// </item>
+		/// <item>
+		/// OuutputFormat.Base64String produces a Base64 encoded string,
+		/// where each group of 3 bytes in the digest is represented by 4 Base64
+		/// characters. For example, a 16-byte digest would be represented by a
+		/// string of 22 Base64 characters.
+		/// </item>
+		/// <item>
+		/// OuutputFormat.PrefixedBase64String produces a Base64 encoded string
+		/// that includes a prefix that identifies the hashing algorithm used to
+		/// generate the digest. The prefix is derived from the name of the
+		/// hashing algorithm, and is followed by an underscore character. For
+		/// example, a 16-byte MD5 digest would be represented by a string of 26
+		/// characters, beginning with "MD5_", and a 64-byte SHA-512 digest
+		/// would be represented by a string of 92 characters, beginning with
+		/// "SHA512_".
+		/// </item>
+		/// </list>
+		/// </param>
+		/// <param name="penmDigestType">
+		/// This parameter specifies the type of the digest, which is used to
+		/// determine the appropriate prefix when the output format is
+		/// PrefixedBase64String. It is of type DigestType, which is an
+		/// enumeration that defines the possible types of digests. The method
+		/// uses  this parameter to look up the corresponding prefix in the
+		/// s_dctHashAlgorithmPrefixes dictionary when the output format is
+		/// PrefixedBase64String. The possible values for this parameter are:
+		/// <list type="number">
+		/// <item>
+		/// DigestType.DIGEST_TYPE_MD5
+		/// </item>
+		/// <item>
+		/// DigestType.DIGEST_TYPE_SHA1
+		/// </item>
+		/// <item>
+		/// DigestType.DIGEST_TYPE_SHA256
+		/// </item>
+		/// <item>
+		/// DigestType.DIGEST_TYPE_SHA384
+		/// </item>
+		/// <item>
+		/// DigestType.DIGEST_TYPE_SHA512
+		/// </item>
+		/// </list>
+		/// </param>
+		/// <returns>
+		/// The returned string is the formatted message digest, based on the
+		/// specified output format. If the output format is
+		/// OuutputFormat.HexadecimalDigits, the returned string will consist of
+		/// hexadecimal characters representing the bytes of the digest. If the
+		/// output format is OuutputFormat.Base64String or
+		/// OuutputFormat.PrefixedBase64String, the output is a Base64 encoded
+		/// string, with or without a prefix that identifies the hashing
+		/// algorithm used to generate the	string. If the output format does
+		/// not match any of the specified cases, it defaults to returning the
+		/// digest as a string of hexadecimal digits.
+		/// </returns>
+		private static string FormatDigestPerOutputFormat ( byte [ ] abytDigest , OuutputFormat penmOutputFormat , DigestType penmDigestType )
+		{
+			switch ( penmOutputFormat )
+			{
+				case OuutputFormat.HexadecimalDigits:
+					return Core.ByteArrayFormatters.ByteArrayToHexDigitString ( abytDigest );
+				case OuutputFormat.Base64String:
+					return Convert.ToBase64String ( abytDigest );
+				case OuutputFormat.PrefixedBase64String:
+					return $"{s_dctHashAlgorithmPrefixes [ penmDigestType ]}{SpecialStrings.UNDERSCORE_CHAR}{Convert.ToBase64String ( abytDigest )}";
+				default:
+					return Core.ByteArrayFormatters.ByteArrayToHexDigitString ( abytDigest );
+			}   // switch ( penmOutputFormat )
+		}   // private static string FormatDigestPerOutputFormat
+
+
+		/// <summary>
+		/// This private method reads the contents of the specified file and
+		/// returns it as a byte array. The method takes a string parameter that
+		/// represents the name of the file to read, which may be a relative or
+		/// absolute path. The method checks if the file name is null and throws
+		/// an ArgumentNullException if it is. It also checks if the file exists
+		/// and throws a FileNotFoundException if it does not. If the file
+		/// exists, the method reads all bytes from the file using the
+		/// File.ReadAllBytes method and returns the byte array containing the
+		/// file contents. This byte array becomes the input to the hashing
+		/// algorithms to compute the message digest.
+		/// </summary>
+		/// <param name="pstrFileName">
+		/// This string represents the name of the file to process, which may be
+		/// a relative or absolute path. The file must exist, and the caller
+		/// must have permission to read it, or an exception will be thrown.
+		/// <para>
+		/// iMPORTANT: Since files are read in binary, they are loaded directly
+		/// into the hash algorithm as byte arrays. This means that the digest
+		/// of a file of ASCII characters and the hash of the file contents read
+		/// into a CLR string object, which is a string of Unicode characters,
+		/// will differ. The reason for this is that the string of Unicode 
+		/// characters yields a different byte stream than the byte stream that
+		/// came from reading the file in binary mode.
+		/// </para>
+		/// </param>
+		/// <returns></returns>
+		/// <exception cref="ArgumentNullException">
+		/// An ArgumentNullException Exception arises when 
+		/// <paramref name="pstrFileName"/> is null.
+		/// </exception>
+		/// <exception cref="FileNotFoundException">
+		/// An FileNotFoundException Exception arises when the file specified by
+		/// <paramref name="pstrFileName"/> cannot be found. A file is reported
+		/// as "not found" when the user does not have permission to read the
+		/// file, or when the file does not exist at the specified or implied
+		/// path.
+		/// </exception>
+		private static byte [ ] GetFileContent ( string pstrFileName )
+		{
+			if ( pstrFileName == null )
+				throw new ArgumentNullException ( nameof ( pstrFileName ) );
+
+			if ( !File.Exists ( pstrFileName ) )
+				throw new FileNotFoundException ( @"Input file not found." , pstrFileName );
+
+			return  File.ReadAllBytes ( pstrFileName );
+		}   // private static byte [ ] GetFileContent
+	}   // class public static class
 }   // partial namespace WizardWrx.Cryptography